Working version with OSC controls

+++ CIMS-J5/src/cims/supervisors/SupervisorMidi_Globals.java
@@ -26,7 +26,7 @@ public class SupervisorMidi_Globals {
 	public static int sDefaultDuration = 250;
 	
 	// Static properties set by external control
-	public static HashMap<String,Float> sActivityWeights;
+	public static HashMap<String,Double> sActivityWeights;
 	public static int sSegmentGap = 3;
 	public static int sRepeatInterval = 0;
 	public static boolean sMetronome = false;
